:recycle: #57 GuestUsersConversationHistoryRepositoryInterface を実装
